requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
blinker==1.6.2
blis==0.7.9
catalogue==1.0.2
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
cymem==2.0.7
Django==4.2.3
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docx2txt==0.8
eventlet==0.33.3
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
kombu==5.3.1
Markdown==3.4.3
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.3
pdfminer.six
phonenumbers==8.13.16
plac==1.1.3
preshed==3.0.8
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.11
PyJWT==2.7.0
pyresparser==1.0.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sortedcontainers==2.4.0
spacy==2.3.5
sqlparse==0.4.4
thinc==7.4.6
tqdm==4.65.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==2.3.6
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
gunicorn